% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\name{ic}
\alias{ic}
\title{Print debugging information with less effort and more clarity}
\usage{
ic(expr)
}
\arguments{
\item{expr}{(optional) An expression to evaluate.}
}
\value{
A value of the passed expression returned invisibly. If no expression was
provided, \code{NULL}.
}
\description{
This function is designed to be used inside the code that is being debugged.
If called without any arguments, prints information on where it was called. If called
with an expression inside, it prints both the expression and its value.
}
\details{
If an expression is given as an argument, it is always evaluated and output is returned.
If no argument is supplied, information about where the function was called is printed:
the file name, the exact line and column number, and the name of the function within which
the function was called. However, this functionality is more sensitive. If the function is
called directly from the console, information about the missing source is displayed.
Finding the file source itself is based on the \link[rlang]{trace_back} function and is
subject to its limitations.
}
\examples{
# using ic() in the console is aimless, it should be used inside other functions
# or when sourcing a file
\donttest{
ic()

foo <- function() {
  2 + 3
  ic()
  3 + 4
  ic()
}

foo()

bar <- function(x) x + 123

ic(bar(765))

a <- ic(234 + 876)

}

}
