% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ic.R
\docType{package}
\name{icecream-package}
\alias{icecream-package}
\alias{icecream}
\title{icecream - Never use print() to debug again}
\description{
Do you ever use print() or log() to debug your code?
Of course you do. IceCream, or ic for short, makes print debugging
a little sweeter. It provides a function that tells you where in your
code it was called or, if an expression is passed to it, what the value
of that expression is.
}
\author{
Dominik Rafacz
}
